feat: strip the leading `v` by default if one exists
This pr changes the behavior slightly when dealing with leading `v`s.
Previously they were handled verbatim. If you had a git tag with
`v1.2.3`, then your version would be `v1.2.3`. Now by default this
strips the leaving `v` so that if you tag `v1.2.3` your version will be
`1.2.3`.

If the user uses the `tagModifier` then this behavior no longer happens
and they user is in full control of the format still.

ref: #51