Expose `calcVcsState` publicly for external usage (#165)
This is necessary in order to make it conditionally used (e.g. only
enabled based on an environment variable), so Mill's own code can take
advantage of https://github.com/com-lihaoyi/mill/pull/4091

As an `Task.Input`, it always is present in the task graph even when the
value is ignored due to an `if` conditional, meaning it always triggers
downstream tasks and tests during selective execution. Whereas as a
helper method I can guard against its usage in an `if` statement (e.g.
checking a `sys.env` variable), and if it isn't used it never ends up in
the build graph at all.

Pull request: https://github.com/lefou/mill-vcs-version/pull/165